restructured fe/be config
